Split //daml-lf/tests:BasicTests into engine and integration tests
