+++ compiler/damlc/tests/daml-test-files/BasicTests.daml
@@ -1,10 +1,13 @@
 -- Copyright (c) 2023 Digital Asset (Switzerland) GmbH and/or its affiliates. All rights reserved.
 -- SPDX-License-Identifier: Apache-2.0
 
+{- DLINT ignore -}
 
+{-# LANGUAGE ApplicativeDo #-}
 
 module BasicTests where
 
+import Daml.Script
 import DA.Time
 import DA.Date
 
@@ -29,7 +32,7 @@ data XYZ2 = XYZ2 { a: XYZ; b: XYZ; c: XYZ }
 testXyz : XYZ2
 testXyz = XYZ2 with a = xyz 1 2 3; b = xyz 2 4 5; c = xyz 3 4 5
 
-test_testXyzTest = scenario do
+test_testXyzTest = script do
   pure testXyz
   pure testXyz
 
@@ -48,13 +51,13 @@ four tok = 2 + (identity 2)
 shadowTest : Int -> Int
 shadowTest x = (\x -> x) $ 2 * x
 
-test_letTest = scenario do
+test_letTest = script do
   let x = four ()
       y = double x
   assertS (x * y == 32)
   pure (x * y)
 
-test_doubleLetTest = scenario do
+test_doubleLetTest = script do
   x <- test_letTest
   assertS (x == 32)
   pure x
@@ -147,7 +150,7 @@ template SimpleMultiParty
     choice HelloMultiParty : Text
       controller p1, p2
       do pure "Hello multi-party"
-    
+
 
 template Composite
    with p1: Party
@@ -162,24 +165,19 @@ template Composite
           pure "hello"
 
 
-ucreate : Party -> Update (ContractId Simple)
-ucreate p =
-  create Simple with p=p
+ccreate : Party -> Commands (ContractId Simple)
+ccreate p =
+  createCmd Simple with p=p
 
-ucreateAndExercise : Party -> Update Text
-ucreateAndExercise p = do
-  cid <- create Simple with p=p
-  exercise cid Hello
+test_screateAndExercise = script do
+  alice <- allocateParty "Alice"
+  cid <- submit alice $ ccreate alice
+  submit alice $ exerciseCmd cid Hello
 
-test_screateAndExercise = scenario do
-  alice <- getParty "Alice"
-  cid <- submit alice $ ucreate alice
-  submit alice $ exercise cid Hello
-
-test_screateAndExerciseComposit = scenario do
-  alice <- getParty "Alice"
-  cid <- submit alice $ create Composite with p1 = alice
-  submit alice $ exercise cid First
+test_screateAndExerciseComposit = script do
+  alice <- allocateParty "Alice"
+  cid <- submit alice $ createCmd Composite with p1 = alice
+  submit alice $ exerciseCmd cid First
 
 
 
@@ -196,40 +194,49 @@ listMatch l =
     h ::_t -> h
     [] -> "<none>"
 
-assertMsgS : Bool -> Text -> Scenario ()
+assertMsgS : Bool -> Text -> Script ()
 assertMsgS b msg =
   if not b then abort msg else return ()
 
-assertS : Bool -> Scenario ()
+assertS : Bool -> Script ()
 assertS b = assertMsgS b ""
 
 assertMsg : Bool -> Text -> Update ()
 assertMsg b msg  =
   if not b then abort msg else return ()
 
-test_listMatchTest = scenario do
+test_listMatchTest : Script ()
+test_listMatchTest = script do
   assertMsgS (listMatch ["x"] == "x") "listMatch [x] == x"
   assertMsgS (listMatch [] == "<none>") "listMatch [] == <none>"
 
-test_sgetTimeTest = scenario do
+test_sgetTimeTest : Script Time
+test_sgetTimeTest = script do
   getTime
 
 
 -- Moves the scenario time forward to the specified time.
-passTo (t: Time) = scenario do
-  now: Time <- pass (days 0)
+passTo : Time -> Script ()
+passTo (t: Time) = script do
+  now: Time <- getTime
   assert {- TODO "passTo can only go forward in time"-} (now < t)
-  pass (subTime t now)
+  passTime (subTime t now)
 
-ugetTime : Update Time
-ugetTime = do
-  t <- getTime
-  pure t
 
-test_ugetTimeTest = scenario do
-  alice <- getParty "Alice"
-  passToDate $ date 2017 Jun 16
-  submit alice $ ugetTime
+template GetTimeHelper
+  with
+    p : Party
+  where
+    signatory p
+    choice GetTimeExe : Time
+      controller p
+      do getTime
+
+test_getTimeTest : Script Time
+test_getTimeTest = script do
+  alice <- allocateParty "Alice"
+  setTime $ datetime 2017 Jun 16 0 0 0
+  submit alice $ createAndExerciseCmd (GetTimeHelper alice) GetTimeExe
 
 
 template TypeWithParameters
@@ -243,25 +250,45 @@ template TypeWithParameters
       controller p
       do pure "hello"
 
-test_typeWithParameters = scenario do
-  alice <- getParty "Alice"
-  submit alice $ create TypeWithParameters with p = alice; v = Some 42
+test_typeWithParameters : Script (ContractId TypeWithParameters)
+test_typeWithParameters = script do
+  alice <- allocateParty "Alice"
+  submit alice $ createCmd TypeWithParameters with p = alice; v = Some 42
 
 
+template RelativeFetchHelper
+  with
+    p : Party
+  where
+    signatory p
+    choice RelativeFetchExe : (ContractId Simple, Simple)
+      controller p
+      do
+        cid <- create Simple with p
+        c <- fetch cid
+        pure (cid, c)
 
-test_createAndFetch = scenario do
-  alice <- getParty "alice"
+template AbsoluteFetchHelper
+  with
+    p : Party
+  where
+    signatory p
+    choice AbsoluteFetchExe : Simple
+      with cid : ContractId Simple
+      controller p
+      do fetch cid
+
+test_createAndFetch : Script Simple
+test_createAndFetch = script do
+  alice <- allocateParty "alice"
   cid <- submit alice $ do
-    cid <- create Simple with p=alice
-    fetch cid -- relative fetch
-    pure cid
+    fst <$> createAndExerciseCmd (RelativeFetchHelper with p=alice) RelativeFetchExe
 
-  -- absolute fetch
   submit alice $ do
-    x <- fetch cid -- absolute fetch
-    create Simple with p=x.p
+    createAndExerciseCmd (AbsoluteFetchHelper with p=alice) (AbsoluteFetchExe cid)
 
-test_exponentiation = scenario do
+test_exponentiation : Script ()
+test_exponentiation = script do
   assertS (3 ^ 5 == 243)
 
 
@@ -295,20 +322,22 @@ template CallablePayout
 
 
 
-test_payoutTest = scenario do
-  alice <- getParty "Alice"
-  bob <- getParty "Bob"
-  clara <- getParty "Clara"
-  pa <- submit alice $ create CallablePayout with giver=alice;receiver=bob
+test_payoutTest : Script (ContractId CallablePayout)
+test_payoutTest = script do
+  alice <- allocateParty "Alice"
+  bob <- allocateParty "Bob"
+  clara <- allocateParty "Clara"
+  pa <- submit alice $ createCmd CallablePayout with giver=alice;receiver=bob
 -- 'Bob' should see all consequences
-  submit bob $ exercise pa Transfer with newReceiver = clara
+  submit bob $ exerciseCmd pa Transfer with newReceiver = clara
 
-test_mustFails = scenario do
-  alice <- getParty "Alice"
-  bob <- getParty "Bob"
-  passToDate $ date 2017 Jun 16
-  cid <- submit alice $ ucreate alice
-  submitMustFail bob $ exercise cid Hello
+test_mustFails : Script ()
+test_mustFails = script do
+  alice <- allocateParty "Alice"
+  bob <- allocateParty "Bob"
+  setTime $ datetime 2017 Jun 16 0 0 0
+  cid <- submit alice $ ccreate alice
+  submitMustFail bob $ exerciseCmd cid Hello
 
 template TwoParties
   with p: Party
@@ -389,14 +418,11 @@ template ExerciseByKey
       do
         exerciseByKey @WithKey (p, 42) SumToK with n = 0
 
-test_failedAuths = scenario do
-  alice <- getParty "alice"
-  bob <- getParty "bob"
-  submitMustFail alice $ (create TwoParties with p=alice; p2=bob)
--- this monster exists solely to test the nesting limit, in the absence of recursive types.
--- generated with
---
---    unlines ["data Nesting" ++ show (n - 1) ++ " = {x: Nesting" ++ show n ++ "}" | n <- [1..150]]
+test_failedAuths : Script ()
+test_failedAuths = script do
+  alice <- allocateParty "alice"
+  bob <- allocateParty "bob"
+  submitMustFail alice $ (createCmd TwoParties with p=alice; p2=bob)
 
 -- Tests for the dynamic computation of fetch actors
 template Fetched
@@ -442,7 +468,10 @@ template Forkable
                 l2 <- exercise c2 Fork with n = n-1
                 pure (l1 ++ l2)
 
-
+-- this monster exists solely to test the nesting limit, in the absence of recursive types.
+-- generated with
+--
+--    unlines ["data Nesting" ++ show (n - 1) ++ " = {x: Nesting" ++ show n ++ "}" | n <- [1..150]]
 data Nesting0 = Nesting0 {x: Nesting1}
 data Nesting1 = Nesting1 {x: Nesting2}
 data Nesting2 = Nesting2 {x: Nesting3}
